Move common clean tasks.
